+++ packages/router/__tests__/middleware.js
@@ -6,15 +6,20 @@ ko.components.register('middleware', {
   template: '<router></router>',
   viewModel: class MiddlewareTest {
     constructor({ t, done }) {
-      Router.use(function*(ctx) {
-        ctx.beforeRenderGlobalMiddlewareHit = true
-        yield
-        ctx.afterRenderGlobalMiddlewareHit = true
-        yield
-        ctx.beforeDisposeGlobalMiddlewareHit = true
-        yield
-        ctx.afterDisposeGlobalMiddlewareHit = true
-      })
+      Router.use((ctx) => ({
+        beforeRender() {
+          ctx.beforeRenderGlobalMiddlewareHit = true
+        },
+        afterRender() {
+          ctx.afterRenderGlobalMiddlewareHit = true
+        },
+        beforeDispose() {
+          ctx.beforeDisposeGlobalMiddlewareHit = true
+        },
+        afterDispose() {
+          ctx.afterDisposeGlobalMiddlewareHit = true
+        }
+      }))
 
       history.replaceState(null, null, '/sync')
 
@@ -29,14 +34,7 @@ ko.components.register('middleware', {
         ],
 
         '/async': [
-          (ctx, _done) => {
-            setTimeout(() => {
-              ctx.waitOver = true
-              _done()
-            }, 200)
-          },
           (ctx) => {
-            t.ok(ctx.waitOver, 'async middleware works with done callback')
             ctx.waitOver = false
             return new Promise((resolve) => {
               setTimeout(() => {
@@ -49,65 +47,11 @@ ko.components.register('middleware', {
             t.ok(ctx.waitOver, 'async middleware works with promise')
             ctx.waitOver = false
 
-            Router.update('/generator')
+            Router.update('/lifecycle')
           }
         ],
 
-        '/generator': [
-          function*(ctx) {
-            t.pass('generator middleware is called')
-
-            t.ok(
-              ctx.beforeRenderGlobalMiddlewareHit,
-              'global middleware before render middleware is executed'
-            )
-            t.ok(
-              ctx.beforeRenderGlobalMiddlewareHit,
-              'route before before render middleware is called after global before render middleware'
-            )
-
-            yield new Promise((resolve) => {
-              setTimeout(() => {
-                ctx.waitOver = true
-                resolve()
-              }, 200)
-            })
-
-            t.ok(
-              ctx.afterRenderGlobalMiddlewareHit,
-              'route after render middleware is called after global after render middleware'
-            )
-
-            Router.update('/object')
-
-            yield
-
-            t.ok(
-              ctx.beforeNavigateHit,
-              'before dispose middleware is called after before navigate callbacks'
-            )
-            t.notOk(
-              ctx.beforeDisposeGlobalMiddlewareHit,
-              'route before dispose middleware is called before global before dispose middleware'
-            )
-
-            yield
-
-            t.notOk(
-              ctx.afterDisposeGlobalMiddlewareHit,
-              'route after dispose middleware is called before global after dispose middleware'
-            )
-          },
-          (ctx) => {
-            t.ok(
-              ctx.waitOver,
-              'generator middleware works with yield-ed promise'
-            )
-          },
-          'generator'
-        ],
-
-        '/object': [
+        '/lifecycle': [
           (ctx) => ({
             beforeRender() {
               t.pass('object middleware is called')
@@ -128,16 +72,18 @@ ko.components.register('middleware', {
                 }, 200)
               })
             },
-            afterRender(done) {
+            afterRender() {
               t.ok(
                 ctx.afterRenderGlobalMiddlewareHit,
                 'route after render middleware is called after global after render middleware'
               )
 
-              setTimeout(() => {
-                ctx.callbackWaitOver = true
-                done()
-              }, 200)
+              return new Promise((resolve) =>
+                setTimeout(() => {
+                  ctx.callbackWaitOver = true
+                  resolve()
+                }, 200)
+              )
             },
             beforeDispose() {
               t.ok(
@@ -171,19 +117,11 @@ ko.components.register('middleware', {
               done()
             }
           }),
-          'object'
+          'lifecycle'
         ]
       })
 
-      ko.components.register('generator', {
-        viewModel: class {
-          constructor(ctx) {
-            ctx.addBeforeNavigateCallback(() => (ctx.beforeNavigateHit = true))
-          }
-        }
-      })
-
-      ko.components.register('object', {
+      ko.components.register('lifecycle', {
         viewModel: class {
           constructor(ctx) {
             ctx.addBeforeNavigateCallback(() => (ctx.beforeNavigateHit = true))