+++ packages/router/src/context.ts
@@ -1,14 +1,8 @@
-import 'core-js/es7/symbol'
 import * as ko from 'knockout'
 import { IContext } from './'
 import { Route } from './route'
-import { Router, Middleware } from './router'
-import {
-  Callback,
-  isPromise,
-  castLifecycleObjectMiddlewareToGenerator,
-  sequence
-} from './utils'
+import { Router, Middleware, LifecycleMiddleware } from './router'
+import { Callback, MaybePromise } from './utils'
 
 export class Context /* implements IContext, use Context & IContext */ {
   public $child?: Context & IContext
@@ -24,8 +18,8 @@ export class Context /* implements IContext, use Context & IContext */ {
 
   private readonly _beforeNavigateCallbacks: Callback<void>[] = []
   private _queue: Promise<void>[] = []
-  private _appMiddlewareDownstream: Callback<void>[] = []
-  private _routeMiddlewareDownstream: Callback<void>[] = []
+  private _appMiddlewareLifecycles: DownstreamLifecycle[] = []
+  private _routeMiddlewareLifecycles: DownstreamLifecycle[] = []
 
   constructor(
     public router: Router,
@@ -132,8 +126,13 @@ export class Context /* implements IContext, use Context & IContext */ {
       callbacks = [...ctx._beforeNavigateCallbacks, ...callbacks]
       ctx = ctx.$child
     }
-    const { success } = await sequence(callbacks)
-    return success
+    for (const cb of callbacks) {
+      const ret = await cb()
+      if (ret === false) {
+        return false
+      }
+    }
+    return true
   }
 
   public render() {
@@ -149,31 +148,14 @@ export class Context /* implements IContext, use Context & IContext */ {
 
   public async runBeforeRender(flush = true) {
     const ctx: Context & IContext = this as any
-    const appMiddlewareDownstream = Context.runMiddleware(
+    this._appMiddlewareLifecycles = await Context.startLifecycle(
       Router.middleware,
       ctx
     )
-    const routeMiddlewareDownstream = Context.runMiddleware(
+    this._routeMiddlewareLifecycles = await Context.startLifecycle(
       this.route.middleware,
       ctx
     )
-
-    const { count: numAppMiddlewareRanPreRedirect } = await sequence(
-      appMiddlewareDownstream
-    )
-    const { count: numRouteMiddlewareRanPreRedirect } = await sequence(
-      routeMiddlewareDownstream
-    )
-
-    this._appMiddlewareDownstream = appMiddlewareDownstream.slice(
-      0,
-      numAppMiddlewareRanPreRedirect
-    )
-    this._routeMiddlewareDownstream = routeMiddlewareDownstream.slice(
-      0,
-      numRouteMiddlewareRanPreRedirect
-    )
-
     if (this.$child && typeof this._redirect === 'undefined') {
       await this.$child.runBeforeRender(false)
     }
@@ -183,10 +165,12 @@ export class Context /* implements IContext, use Context & IContext */ {
   }
 
   public async runAfterRender() {
-    await sequence([
-      ...this._appMiddlewareDownstream,
-      ...this._routeMiddlewareDownstream
-    ])
+    for (const l of [
+      ...this._appMiddlewareLifecycles,
+      ...this._routeMiddlewareLifecycles
+    ]) {
+      if (l.afterRender) await l.afterRender()
+    }
     await this.flushQueue()
   }
 
@@ -194,10 +178,12 @@ export class Context /* implements IContext, use Context & IContext */ {
     if (this.$child && typeof this._redirect === 'undefined') {
       await this.$child.runBeforeDispose(false)
     }
-    await sequence([
-      ...this._routeMiddlewareDownstream,
-      ...this._appMiddlewareDownstream
-    ])
+    for (const l of [
+      ...this._routeMiddlewareLifecycles,
+      ...this._appMiddlewareLifecycles
+    ]) {
+      if (l.beforeDispose) await l.beforeDispose()
+    }
     if (flush) {
       await this.flushQueue()
     }
@@ -207,10 +193,12 @@ export class Context /* implements IContext, use Context & IContext */ {
     if (this.$child && typeof this._redirect === 'undefined') {
       await this.$child.runAfterDispose(false)
     }
-    await sequence([
-      ...this._routeMiddlewareDownstream,
-      ...this._appMiddlewareDownstream
-    ])
+    for (const l of [
+      ...this._routeMiddlewareLifecycles,
+      ...this._appMiddlewareLifecycles
+    ]) {
+      if (l.afterDispose) await l.afterDispose()
+    }
     if (flush) {
       await this.flushQueue()
     }
@@ -224,31 +212,29 @@ export class Context /* implements IContext, use Context & IContext */ {
     await Promise.all<Promise<void>>([thisQueue, ...childQueues])
   }
 
-  private static runMiddleware(
+  private static async startLifecycle(
     middleware: Middleware[],
     ctx: Context & IContext
-  ): Callback<void>[] {
-    return middleware.map((fn) => {
-      const runner = castLifecycleObjectMiddlewareToGenerator(fn)(ctx)
-      let beforeRender = true
-      return async () => {
-        const ret = runner.next()
-        if (isPromise(ret)) {
-          await ret
-        } else if (
-          isPromise((ret as IteratorResult<Promise<void> | void>).value)
-        ) {
-          await (ret as IteratorResult<Promise<void> | void>).value
-        }
-        if (beforeRender) {
-          // this should only block the sequence for the first call,
-          // and allow cleanup after the redirect
-          beforeRender = false
-          return typeof ctx._redirect === 'undefined'
-        } else {
-          return true
-        }
+  ): Promise<DownstreamLifecycle[]> {
+    const downstream: DownstreamLifecycle[] = []
+
+    for (const fn of middleware) {
+      if (typeof ctx._redirect !== 'undefined') {
+        break
       }
-    })
+      const lifecycle = await fn(ctx)
+      if (lifecycle) {
+        if (lifecycle.beforeRender) await lifecycle.beforeRender()
+        downstream.push(lifecycle)
+      }
+    }
+
+    return downstream
   }
+}
+
+type DownstreamLifecycle = {
+  afterRender?(): MaybePromise<void>
+  beforeDispose?(): MaybePromise<void>
+  afterDispose?(): MaybePromise<void>
 }