+++ packages/router/CHANGELOG.md
@@ -16,10 +16,16 @@ For now it is published as `next`
 
 - Delay navigation complete until `descendantsComplete` (see [Knockout 3.5.0-beta][], [Knockout #2319][])
 
+### Removed
+
+- Generator middleware (no longer requires `Symbol.asyncIterator`) support
+- Async via `done` callbacks in middleware and beforeNavigate callbacks (use promises)
+
 ### Fixed
 
 - Static `Router.isNavigating` shouldn't throw if read before initialization
 - Use strict TypeScript checking (this is relevant if you use TypeScript with strict to consume, as the types declaration points to the source)
+- Issue with lifecycle middleware not being cast to generator correctly (generators are gone, internal implementation uses the lifecycle)
 
 [knockout 3.5.0-beta]: https://github.com/knockout/knockout/releases/tag/v3.5.0-beta
 [knockout #2319]: https://github.com/knockout/knockout/pull/2319